get all jobs updated
