+++ routes/postJobRoutes.js
@@ -7,10 +7,8 @@ import {
   getAllJobs,
   deleteJob,
   updateJob,
-  getAllAvailableJobs
 } from "../controllers/postJobController.js";
 
-import { companyDetails } from "../controllers/companyController.js";
 import postJobValidator from "../Validators/postJobValidator.js";
 import { uploadLogoMiddleware, verifyCompanyToken } from "../middleware/companyMiddleware.js";
 import { getCompanyAllInfo } from "../controllers/companyController.js";
@@ -26,9 +24,8 @@ router.post(
 
 router.get("/getOne/:id", getOneJob);
 router.get("/getAll/:id", getAllJobs);
-router.get("/availableJobs", getAllAvailableJobs);
 router.delete("/deleteJob/:id", deleteJob);
 router.put("/update/:id", uploadLogoMiddleware("").none(), verifyCompanyToken, updateJob);
-// router.get("/companyDetails/:id", companyDetails);
+
 
 export default router;
